+++ app/src/main/java/com/example/criptomonedasapp/model/network/FeesModel.kt
@@ -1,0 +1,8 @@
+package com.example.criptomonedasapp.model.network
+
+
+
+data class FeesModel(
+    var flat_rate: FlatRateModel,
+    var structure: List<StructureModel>,
+)
