+++ app/src/main/java/com/example/criptomonedasapp/model/network/FlatRateModel.kt
@@ -1,0 +1,6 @@
+package com.example.criptomonedasapp.model.network
+
+data class FlatRateModel(
+    var maker: String,
+    var taker: String,
+)
