+++ app/src/main/java/com/example/criptomonedasapp/model/network/RollingAverageChangeModel.kt
@@ -1,0 +1,7 @@
+package com.example.criptomonedasapp.model.network
+
+import com.google.gson.annotations.SerializedName
+
+data class RollingAverageChangeModel(
+    @SerializedName("6") var rolling_average_change: String,
+)
