+++ app/src/main/java/com/example/criptomonedasapp/model/network/StructureModel.kt
@@ -1,0 +1,7 @@
+package com.example.criptomonedasapp.model.network
+
+data class StructureModel(
+    var volume :String,
+    var maker :String,
+    var taker :String,
+)
