+++ app/src/main/java/com/example/criptomonedasapp/model/network/response/CoinsResponse.kt
@@ -1,0 +1,13 @@
+package com.example.criptomonedasapp.model.network.response
+
+import com.example.criptomonedasapp.model.network.CoinListModel
+import com.google.gson.annotations.SerializedName
+
+data class CoinsResponse (
+    var success: String,
+    @SerializedName("payload") var CoinList: List<CoinListModel>
+)
+
+
+
+
