+++ app/src/main/java/com/example/criptomonedasapp/model/network/response/TickerResponseModel.kt
@@ -1,6 +1,6 @@
-package com.example.criptomonedasapp.model.response
+package com.example.criptomonedasapp.model.network.response
 
-import com.example.criptomonedasapp.model.CoinDetailModel
+import com.example.criptomonedasapp.model.network.CoinDetailModel
 import com.google.gson.annotations.SerializedName
 
 data class TickerResponseModel (
