+++ app/src/main/java/com/example/criptomonedasapp/mvvm/iu/coindetailfragment/CoinDetailViewModel.kt
@@ -1,11 +1,11 @@
-package com.example.criptomonedasapp.mvvm.detailscoinfragment
+package com.example.criptomonedasapp.mvvm.iu.coindetailfragment
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.criptomonedasapp.repository.CryptocurrenciesRepository
-import com.example.criptomonedasapp.model.CoinDetailModel
+import com.example.criptomonedasapp.model.network.CoinDetailModel
+import com.example.criptomonedasapp.mvvm.data.repository.CryptocurrenciesRepositoryImpl
 import kotlinx.coroutines.Dispatchers
 
 import kotlinx.coroutines.launch
@@ -13,10 +13,8 @@ import kotlinx.coroutines.withContext
 
 class CoinDetailViewModel : ViewModel() {
 
-    var  repository : CryptocurrenciesRepository = CryptocurrenciesRepository()
+    var  repository : CryptocurrenciesRepositoryImpl = CryptocurrenciesRepositoryImpl()
     private lateinit var list: CoinDetailModel
-
-
 
     private val _detail = MutableLiveData<CoinDetailModel>()
     val detail: LiveData<CoinDetailModel> = _detail
