+++ app/src/main/java/com/example/criptomonedasapp/services/APIService.kt
@@ -1,8 +1,8 @@
 package com.example.criptomonedasapp.services
 
 
-import com.example.criptomonedasapp.model.response.CoinsResponse
-import com.example.criptomonedasapp.model.response.TickerResponseModel
+import com.example.criptomonedasapp.model.network.response.CoinsResponse
+import com.example.criptomonedasapp.model.network.response.TickerResponseModel
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
