+++ app/src/main/res/layout/coin_detail_fragment.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:padding="16dp"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".mvvm.detailscoinfragment.CoinDetailFragment">
+    tools:context=".mvvm.iu.coindetailfragment.CoinDetailFragment">
 
     <ImageView
         android:id="@+id/imgIconCoin"
