+++ app/src/main/res/layout/coin_list_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".mvvm.listcoinsfragment.ListCoinsFragment">
+    tools:context=".mvvm.iu.listcoinsfragment.ListCoinsFragment">
 
    <androidx.recyclerview.widget.RecyclerView
        android:id="@+id/rvListCoins"
